fix(driver/bpf): fixed bpf compilation problem on kernel >5.0
